backend/requirements.txt
fastapi
uvicorn[standard]
numpy
pillow
opencv-python
python-multipart
tensorflow-cpu>=2.15
